Helpers, Killers, and Pointers
REMOVED
- Got rid of unneccessary attribute declarations at the top of Board.cpp
and Species.cpp. These weren't static, so they could just be
declared/initialized in the header files.
- Removed the "dead cell" functionality from Board::NeighborBehavior().

FIXED
- Changed incorrect calls to the Species::killedBy and Species::helpedBy
vectors in the Species::AddKiller() and Species::AddHelper() functions
so that they assign to the corresponding vector directly instead of
attempting to go through the Species::GetKillers() and
Species::GetHelpers() function.
- Checking a living cell's helpers and killers will no longer make an
incorrect comparison between the bothCheck vector value and the cell's
helper/killer IDs. It will instead compare the neighbor IDs (found
through the bothCheck vector) and the helper/killer IDs. Neighbors and
Killers should now be acting as expected.

CHANGED
-Renamed Board::NeighborBehavior and Board::NeighborConflict() to
LivingBehavior() and DeadBehavior(), respectively, to better reflect
what they do.
- Changed how Board::DeadBehavior() checks which Species (if any)
reproduces onto the cell. The previous implementation assumed that the
Species (singular or plural) that had the most neighbors to the dead
cell would be the only one(s) able to reproduce onto the cell (which is
incorrect).
- Switched many instances of Species to be Species pointers. This gave a
considerable speed boost.